docs: document API classes
