+++ api/build.gradle.kts
@@ -1,0 +1,17 @@
+plugins {
+    id("java-convention")
+    `maven-publish`
+}
+
+java {
+    withSourcesJar()
+    withJavadocJar()
+}
+
+publishing {
+    publications {
+        create<MavenPublication>("mavenJava") {
+            from(components["java"])
+        }
+    }
+}
