+++ api/src/main/java/com/turikhay/mc/mapmodcompanion/LevelMapProperties.java
@@ -2,10 +2,18 @@ package com.turikhay.mc.mapmodcompanion;
 
 import java.io.*;
 
+/**
+ * Binary format used by VoxelMap to communicate the world id.
+ */
 public interface LevelMapProperties {
+
+    /**
+     * Deserializes standard id packets.
+     */
     class Deserializer implements Id.Deserializer<StandardId> {
         private static Deserializer INSTANCE;
 
+        /** {@inheritDoc} */
         @Override
         public StandardId deserialize(byte[] data) throws MalformedPacketException {
             DataInputStream in = new DataInputStream(new ByteArrayInputStream(data));
@@ -20,19 +28,32 @@ public interface LevelMapProperties {
             }
         }
 
+        /**
+         * Returns a shared instance of the deserializer.
+         */
         public static Deserializer instance() {
             return INSTANCE == null ? INSTANCE = new Deserializer() : INSTANCE;
         }
     }
 
+    /**
+     * Serializes standard ids into packet byte arrays.
+     */
     class Serializer implements Id.Serializer<StandardId> {
         private static Serializer INSTANCE;
 
+        /** {@inheritDoc} */
         @Override
         public byte[] serialize(StandardId id) {
             return serialize(id.getId());
         }
 
+        /**
+         * Encodes the supplied id into the binary representation.
+         *
+         * @param id numeric world id
+         * @return encoded packet bytes
+         */
         public byte[] serialize(int id) {
             ByteArrayOutputStream array = new ByteArrayOutputStream();
             try(DataOutputStream out = new DataOutputStream(array)) {
@@ -44,6 +65,9 @@ public interface LevelMapProperties {
             return array.toByteArray();
         }
 
+        /**
+         * Returns a shared instance of the serializer.
+         */
         public static Serializer instance() {
             return INSTANCE == null ? INSTANCE = new Serializer() : INSTANCE;
         }
