+++ api/src/main/java/com/turikhay/mc/mapmodcompanion/PrefixedId.java
@@ -4,36 +4,59 @@ import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.util.Objects;
 
+/**
+ * Identifier encoded with a zero-prefixed length and optional magic marker.
+ * <p>
+ * The format matches the one used by Xaero's minimap when responding to world
+ * id requests. The {@link Deserializer} and {@link Serializer} nested classes
+ * can be used to convert between the packet representation and this class.
+ */
 public class PrefixedId implements Id {
     private final int padding;
     private final boolean usesMagicByte;
     private final int id;
 
+    /**
+     * Creates a new prefixed id.
+     *
+     * @param padding       number of leading zero bytes before the marker
+     * @param usesMagicByte whether the {@link #MAGIC_MARKER} is present
+     * @param id            numeric world id
+     */
     public PrefixedId(int padding, boolean usesMagicByte, int id) {
         this.padding = padding;
         this.usesMagicByte = usesMagicByte;
         this.id = id;
     }
 
+    /**
+     * Deprecated constructor that assumes the packet uses the magic marker.
+     */
     @Deprecated
     public PrefixedId(int padding, int id) {
         this(padding, true, id);
     }
 
+    /** {@inheritDoc} */
     @Override
     public int getId() {
         return id;
     }
 
+    /**
+     * Number of zero bytes prepended to the packet.
+     */
     public int getPadding() {
         return padding;
     }
 
+    /** {@inheritDoc} */
     @Override
     public PrefixedId withIdUnchecked(int id) {
         return new PrefixedId(this.padding, this.usesMagicByte, id);
     }
 
+    /** {@inheritDoc} */
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
@@ -42,11 +65,13 @@ public class PrefixedId implements Id {
         return padding == that.padding && usesMagicByte == that.usesMagicByte && id == that.id;
     }
 
+    /** {@inheritDoc} */
     @Override
     public int hashCode() {
         return Objects.hash(padding, usesMagicByte, id);
     }
 
+    /** {@inheritDoc} */
     @Override
     public String toString() {
         return "PrefixedId{" +
@@ -56,9 +81,17 @@ public class PrefixedId implements Id {
                 '}';
     }
 
+    /**
+     * Deserializes prefixed id packets.
+     *
+     * <pre>{@code
+     * PrefixedId id = PrefixedId.Deserializer.instance().deserialize(data);
+     * }</pre>
+     */
     public static class Deserializer implements Id.Deserializer<PrefixedId> {
         private static Deserializer INSTANCE;
 
+        /** {@inheritDoc} */
         @Override
         public PrefixedId deserialize(byte[] data) throws MalformedPacketException {
             DataInputStream in = new DataInputStream(new ByteArrayInputStream(data));
@@ -101,14 +134,21 @@ public class PrefixedId implements Id {
             }
         }
 
+        /**
+         * Returns a shared instance of the deserializer.
+         */
         public static Deserializer instance() {
             return INSTANCE == null ? INSTANCE = new Deserializer() : INSTANCE;
         }
     }
 
+    /**
+     * Serializes {@link PrefixedId} instances into packet byte arrays.
+     */
     public static class Serializer implements Id.Serializer<PrefixedId> {
         private static Serializer INSTANCE;
 
+        /** {@inheritDoc} */
         @Override
         public byte[] serialize(PrefixedId id) {
             ByteArrayOutputStream array = new ByteArrayOutputStream();
@@ -128,6 +168,9 @@ public class PrefixedId implements Id {
             return array.toByteArray();
         }
 
+        /**
+         * Returns a shared instance of the serializer.
+         */
         public static Serializer instance() {
             return INSTANCE == null ? INSTANCE = new Serializer() : INSTANCE;
         }
