+++ api/src/main/java/com/turikhay/mc/mapmodcompanion/package-info.java
@@ -1,3 +1,6 @@
+/**
+ * Public API exposing helper classes for querying world identifiers.
+ */
 @ParametersAreNonnullByDefault
 @ReturnValuesAreNonnullByDefault
 package com.turikhay.mc.mapmodcompanion;
