+++ bungee/build.gradle.kts
@@ -29,6 +29,7 @@ tasks {
 }
 
 dependencies {
+    implementation(project(":api"))
     implementation(project(":common"))
     implementation(libs.bstats.bungeecord)
     compileOnly(libs.bungeecord.api)
