+++ common/build.gradle.kts
@@ -1,3 +1,7 @@
 plugins {
     id("java-convention")
 }
+
+dependencies {
+    implementation(project(":api"))
+}
