--- common/src/main/java/com/turikhay/mc/mapmodcompanion/Channels.java
@@ -1,8 +1,0 @@
-package com.turikhay.mc.mapmodcompanion;
-
-public interface Channels {
-    String XAERO_MINIMAP_CHANNEL = "xaerominimap:main";
-    String XAERO_WORLDMAP_CHANNEL = "xaeroworldmap:main";
-    String WORLDID_CHANNEL = "worldinfo:world_id";
-    String WORLDID_LEGACY_CHANNEL = "world_id";
-}
