--- common/src/main/java/com/turikhay/mc/mapmodcompanion/LightweightException.java
@@ -1,13 +1,0 @@
-package com.turikhay.mc.mapmodcompanion;
-
-import javax.annotation.Nullable;
-
-public class LightweightException extends Exception {
-    public LightweightException(String message, @Nullable Throwable cause) {
-        super(message, cause, false, false);
-    }
-
-    public LightweightException(String message) {
-        this(message, null);
-    }
-}
