+++ settings.gradle.kts
@@ -1,6 +1,7 @@
 rootProject.name = "mapmodcompanion"
 
 include(":common")
+include(":api")
 include(":bungee")
 include(":spigot")
 include(":velocity")
