+++ spigot/build.gradle.kts
@@ -39,6 +39,7 @@ val spigot_version: String by project
 val protocolLib_version: String by project
 
 dependencies {
+    implementation(project(":api"))
     implementation(project(":common"))
     implementation(libs.bstats.bukkit)
